+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a CHANGELOG](https://keepachangelog.com/).
 
+## [6.1.3]
+
+### Added
+
+- Nationbuilder now supports tags.
+- Range slider now supports custom input value with single submit option.
+
+### Fixed
+
+- Disabled button in admin.
+
 ## [6.1.2]
 
 ### Added
@@ -1102,6 +1113,7 @@ This projects adheres to [Semantic Versioning](https://semver.org/) and [Keep a 
 
 - Initial production release.
 
+[6.1.3]: https://github.com/infinum/eightshift-forms/compare/6.1.2...6.1.3
 [6.1.2]: https://github.com/infinum/eightshift-forms/compare/6.1.1...6.1.2
 [6.1.1]: https://github.com/infinum/eightshift-forms/compare/6.1.0...6.1.1
 [6.1.0]: https://github.com/infinum/eightshift-forms/compare/6.0.0...6.1.0